# docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: LLM Platform API
  description: |
    Enterprise-grade Large Language Model API platform providing OpenAI-compatible endpoints
    for chat completions, embeddings, function calling, and more.
    
    ## Features
    - OpenAI-compatible API endpoints
    - Multiple LLM model support (GPT, Claude, Llama, etc.)
    - Streaming responses
    - Function calling capabilities
    - Embeddings and semantic search
    - Enterprise security and authentication
    - Rate limiting and usage tracking
    - Fine-tuning and model management
    
    ## Authentication
    All API requests require authentication using API keys. Include your API key in the
    Authorization header:
    ```
    Authorization: Bearer your_api_key_here
    ```
    
    ## Rate Limits
    Rate limits are enforced per API key:
    - Free tier: 20 requests/minute
    - Pro tier: 3,500 requests/minute  
    - Enterprise: Custom limits
    
    ## Error Handling
    The API uses conventional HTTP response codes and returns detailed error information
    in JSON format.
  version: 1.0.0
  contact:
    name: LLM Platform Team
    email: support@llm-platform.com
    url: https://docs.llm-platform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://llm-platform.com/terms
servers:
  - url: https://api.llm-platform.com
    description: Production server
  - url: https://staging-api.llm-platform.com
    description: Staging server
...
security:
  - BearerAuth: []
paths:
...
  /v1/chat/completions:
    post:
      summary: Create chat completion
      description: |
        Creates a model response for the given chat conversation.
        Supports both streaming and non-streaming responses.
      tags: [Chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
            examples:
              simple_chat:
                summary: Simple chat completion
                value:
                  model: gpt-3.5-turbo
                  messages:
                    - role: user
                      content: "Hello, how are you?"
                  max_tokens: 150
              function_calling:
                summary: Function calling example
                value:
                  model: gpt-3.5-turbo
                  messages:
                    - role: user
                      content: "What's the weather like in San Francisco?"
                  functions:
                    - name: get_weather
                      description: Get current weather for a location
                      parameters:
                        type: object
                        properties:
                          location:
                            type: string
                            description: The city and state
                        required: [location]
              streaming:
                summary: Streaming response
                value:
                  model: gpt-3.5-turbo
                  messages:
                    - role: user
                      content: "Write a short story"
                  max_tokens: 300
                  stream: true
      responses:
        '200':
          description: Chat completion response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
            text/event-stream:
              schema:
                type: string
                description: Server-sent events for streaming responses
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimit'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        API key authentication. Include your API key in the Authorization header:
        `Authorization: Bearer your_api_key_here`
  schemas:
...
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant, function]
          description: The role of the message author
        content:
          type: string
          nullable: true
          description: The contents of the message
        name:
          type: string
          description: The name of the author of this message
        function_call:
          type: object
          description: Function call information
          properties:
            name:
              type: string
              description: Function name
            arguments:
              type: string
              description: Function arguments as JSON string
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          description: ID of the model to use
          example: gpt-3.5-turbo
        messages:
          type: array
          description: List of messages comprising the conversation
          items:
            $ref: '#/components/schemas/ChatMessage'
          minItems: 1
        functions:
          type: array
          description: List of functions the model may call
          items:
            $ref: '#/components/schemas/FunctionDefinition'
        function_call:
          oneOf:
            - type: string
              enum: [none, auto]
            - type: object
              properties:
                name:
                  type: string
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 1
          description: Sampling temperature
        top_p:
          type: number
          minimum: 0
          maximum: 1
          default: 1
          description: Nucleus sampling parameter
        n:
          type: integer
          minimum: 1
          maximum: 128
          default: 1
          description: Number of completions to generate
        stream:
          type: boolean
          default: false
          description: Whether to stream partial results
        stop:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
              maxItems: 4
          description: Stop sequences
        max_tokens:
          type: integer
          minimum: 1
          description: Maximum number of tokens to generate
        presence_penalty:
          type: number
          minimum: -2
          maximum: 2
          default: 0
        frequency_penalty:
          type: number
          minimum: -2
          maximum: 2
          default: 0
        logit_bias:
          type: object
          additionalProperties:
            type: number
            minimum: -100
            maximum: 100
        user:
          type: string
          description: Unique identifier for the end-user
    ChatCompletionResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the completion
        object:
          type: string
          example: chat.completion
        created:
          type: integer
          description: Unix timestamp of completion creation
        model:
          type: string
          description: Model used for completion
        choices:
          type: array
          items:
            $ref: '#/components/schemas/ChatChoice'
        usage:
          $ref: '#/components/schemas/Usage'
    ChatChoice:
      type: object
      properties:
        index:
          type: integer
          description: Choice index
        message:
          $ref: '#/components/schemas/ChatMessage'
        finish_reason:
          type: string
          enum: [stop, length, function_call, content_filter]
          description: Reason why the completion finished
    FunctionDefinition:
      type: object
      required: [name, description, parameters]
      properties:
        name:
          type: string
          description: Function name
        description:
          type: string
          description: Function description
        parameters:
          type: object
          description: Function parameters as JSON Schema
    Usage:
      type: object
      properties:
        prompt_tokens:
          type: integer
          description: Number of tokens in the prompt
        completion_tokens:
          type: integer
          description: Number of tokens in the completion
        total_tokens:
          type: integer
          description: Total number of tokens used
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
              description: Human-readable error message
            type:
              type: string
              description: Error type
            param:
              type: string
              nullable: true
              description: Parameter that caused the error
            code:
              type: string
              nullable: true
              description: Error code
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              message: "Invalid request format"
              type: "invalid_request_error"
              param: null
              code: null
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              message: "Invalid API key provided"
              type: "invalid_request_error"
              param: null
              code: "invalid_api_key"
    RateLimit:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              message: "Rate limit exceeded"
              type: "rate_limit_error"
              param: null
              code: "rate_limit_exceeded"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              message: "Internal server error"
              type: "server_error"
              param: null
              code: "internal_error"
...