# fastapi_service/app/api/v2/endpoints/llm_endpoints.py
"""
Enhanced LLM Endpoints for API v2
Includes function calling, advanced features, and improved responses
"""

import uuid
import time
from fastapi import APIRouter, HTTPException, Depends, Request
from typing import Dict, Any
import logging

from ..models import (
    EnhancedChatRequest, EnhancedChatResponse, EnhancedModelInfo,
    BatchRequest, BatchResponse, EnhancedUsage, EnhancedChatChoice,
    EnhancedMessage, MessageRole
)

logger = logging.getLogger(__name__)

router = APIRouter()


@router.post("/chat/completions", response_model=EnhancedChatResponse)
async def create_enhanced_chat_completion(
    request: EnhancedChatRequest
) -> EnhancedChatResponse:
    """
    Create enhanced chat completion with function calling and advanced features
    """
    try:
        completion_id = f"chatcmpl-v2-{uuid.uuid4().hex[:16]}"
        
        # Enhanced mock response with function calling support
        user_messages = [msg for msg in request.messages if msg.role == MessageRole.USER]
        last_message = user_messages[-1].content if user_messages else "Hello"
        
        # Check if tools are available and simulate function calling
        response_message = EnhancedMessage(
            role=MessageRole.ASSISTANT,
            content=f"Enhanced v2 response to: '{last_message[:50]}...' with advanced features enabled"
        )
        
        # If tools are provided, simulate tool usage
        if request.tools:
            response_message.tool_calls = [{
                "id": f"call_{uuid.uuid4().hex[:8]}",
                "type": "function",
                "function": {
                    "name": request.tools[0].function.name,
                    "arguments": '{"query": "example"}'
                }
            }]
        
        # Enhanced usage tracking
        prompt_tokens = sum(len(msg.content.split()) if msg.content else 0 for msg in request.messages)
        completion_tokens = len(response_message.content.split()) if response_message.content else 0
        
        return EnhancedChatResponse(
            id=completion_id,
            object="chat.completion",
            created=int(time.time()),
            model=request.model,
            choices=[
                EnhancedChatChoice(
                    index=0,
                    message=response_message,
                    finish_reason="stop" if not request.tools else "tool_calls"
                )
            ],
            usage=EnhancedUsage(
                prompt_tokens=prompt_tokens,
                completion_tokens=completion_tokens,
                total_tokens=prompt_tokens + completion_tokens,
                cost=0.002 * (prompt_tokens + completion_tokens),  # Mock cost
                processing_time_ms=150.5,
                cache_hit=False
            ),
            provider=request.provider.value,
            system_fingerprint="fp_v2_enhanced",
            metadata={
                "request_id": completion_id,
                "api_version": "2.0",
                "features_used": ["enhanced_chat", "function_calling"] if request.tools else ["enhanced_chat"]
            }
        )
        
    except Exception as e:
        logger.error(f"Error in enhanced chat completion: {e}")
        raise HTTPException(status_code=500, detail="Internal server error")


@router.get("/models/enhanced", response_model=list[EnhancedModelInfo])
async def list_enhanced_models() -> list[EnhancedModelInfo]:
    """
    List enhanced models with detailed capabilities and pricing
    """
    try:
        enhanced_models = [
            EnhancedModelInfo(
                id="gpt-4-turbo",
                object="model",
                created=int(time.time()),
                owned_by="openai",
                provider="openai",
                capabilities=["chat", "function_calling", "vision", "reasoning"],
                context_length=128000,
                max_output_tokens=4096,
                description="GPT-4 Turbo with enhanced capabilities",
                pricing={"input_tokens": 0.01, "output_tokens": 0.03},
                performance_metrics={"latency_p95": 2.5, "throughput": 1000},
                supported_languages=["en", "es", "fr", "de", "zh"],
                version="2024-01-25",
                status="active"
            ),
            EnhancedModelInfo(
                id="claude-3-opus",
                object="model",
                created=int(time.time()),
                owned_by="anthropic",
                provider="anthropic",
                capabilities=["chat", "reasoning", "code_generation"],
                context_length=200000,
                max_output_tokens=4096,
                description="Claude 3 Opus with advanced reasoning",
                pricing={"input_tokens": 0.015, "output_tokens": 0.075},
                performance_metrics={"latency_p95": 3.2, "throughput": 800},
                supported_languages=["en", "es", "fr", "de", "zh", "ja"],
                version="2024-02-29",
                status="active"
            )
        ]
        
        return enhanced_models
        
    except Exception as e:
        logger.error(f"Error listing enhanced models: {e}")
        raise HTTPException(status_code=500, detail="Internal server error")